Update to Vaadin 7.2 beta1; recompiling widgetset and adapting Responsive settings

+++ src/main/java/com/pathfinder/util/widgetset/client/datetime/DateTimeConnector.java
@@ -36,7 +36,7 @@ public class DateTimeConnector extends AbstractComponentConnector {
 
 	private final static int CHECK_IF_TIME_TO_GO_HOME_INTERVALL = 5 * 1000;
 
-	private long timeOld = 0;
+	private double timeOld = 0;
 
 	public DateTimeConnector() {
